Restructured files into folders
