+++ main.py
@@ -5,10 +5,10 @@ Application
 
 from announcer import TripAnnouncer
 from database import LocalDatabase
-from location_specifier import CommandlineLocationUpdator
+from io.location_specifier import CommandlineLocationUpdator
 from models import RouteLocation
 from stops_finder import NextStopsFinder
-from viewer import CommandlineDisplay
+from io.viewer import CommandlineDisplay
 
 
 def main() -> None:
